Rename modules to use snake_case (#1498)
